[Build_Sys_Refact] Stable phase for Makefile refactory
- Scalable approach introduced
- BUILD_ARTIFACT_NAME allows us to select the desired driver to build
- Since it is not a merged applicative software, drivers are intended to
  be built separately
- Test section is added as a placeholder
- drv_dac folder and file has been added as a placeholder

This fourth phase for Makefile refactory (stable) allows us to introduce an scalable and
maintainable approach for each of the drivers desired to be built. Less
details to improve.

By selecting in the BUILD_ARTIFACT_NAME the driver, build process is
executed successfully
Target "flash" was introduced in order to flash in the Discovery board
the built image.